Update Makefile and test vector generation
