+++ tokens/token-test-vectors.json
@@ -1,0 +1,1 @@
+[{"token_type":2,"issuer_name":"6973737565722e6578616d706c65","redemption_context":"619e580a1fe8a7a8938fb3c007af1f6c15f3a94037054fb4f3a7e20bdc15e810","origin_info":"6f726967696e2e6578616d706c65","nonce":"f93a21c15f70a2e67f5efa2f9a5b504c8c8ef895c3f61f31f4abb3a1dc5e2b60","token_key_id":"ca572f8982a9ca248a3056186322d93ca147266121ddeb5632c07f1f71cd2708","token_authenticator_input":"0002f93a21c15f70a2e67f5efa2f9a5b504c8c8ef895c3f61f31f4abb3a1dc5e2b6051024ebd20f96a17d999b20a4ff8f6e9bbd72f61ee0aa8b0d3a50b3f4979439aca572f8982a9ca248a3056186322d93ca147266121ddeb5632c07f1f71cd2708"},{"token_type":2,"issuer_name":"6973737565722e6578616d706c65","redemption_context":"","origin_info":"6f726967696e2e6578616d706c65","nonce":"f93a21c15f70a2e67f5efa2f9a5b504c8c8ef895c3f61f31f4abb3a1dc5e2b60","token_key_id":"ca572f8982a9ca248a3056186322d93ca147266121ddeb5632c07f1f71cd2708","token_authenticator_input":"0002f93a21c15f70a2e67f5efa2f9a5b504c8c8ef895c3f61f31f4abb3a1dc5e2b6011e15c91a7c2ad02abd66645802373db1d823bea80f08d452541fb2b62b5898bca572f8982a9ca248a3056186322d93ca147266121ddeb5632c07f1f71cd2708"},{"token_type":2,"issuer_name":"6973737565722e6578616d706c65","redemption_context":"","origin_info":"","nonce":"f93a21c15f70a2e67f5efa2f9a5b504c8c8ef895c3f61f31f4abb3a1dc5e2b60","token_key_id":"ca572f8982a9ca248a3056186322d93ca147266121ddeb5632c07f1f71cd2708","token_authenticator_input":"0002f93a21c15f70a2e67f5efa2f9a5b504c8c8ef895c3f61f31f4abb3a1dc5e2b60b741ec1b6fd05f1e95f8982906aec1612896d9ca97d53eef94ad3c9fe023f7a4ca572f8982a9ca248a3056186322d93ca147266121ddeb5632c07f1f71cd2708"},{"token_type":2,"issuer_name":"6973737565722e6578616d706c65","redemption_context":"619e580a1fe8a7a8938fb3c007af1f6c15f3a94037054fb4f3a7e20bdc15e810","origin_info":"","nonce":"f93a21c15f70a2e67f5efa2f9a5b504c8c8ef895c3f61f31f4abb3a1dc5e2b60","token_key_id":"ca572f8982a9ca248a3056186322d93ca147266121ddeb5632c07f1f71cd2708","token_authenticator_input":"0002f93a21c15f70a2e67f5efa2f9a5b504c8c8ef895c3f61f31f4abb3a1dc5e2b60d703248fbfcd407d48f9e972d2a7770c2a9d7c97f5449ba4863f968c7d0fa993ca572f8982a9ca248a3056186322d93ca147266121ddeb5632c07f1f71cd2708"},{"token_type":2,"issuer_name":"6973737565722e6578616d706c65","redemption_context":"619e580a1fe8a7a8938fb3c007af1f6c15f3a94037054fb4f3a7e20bdc15e810","origin_info":"6f726967696e2e6578616d706c652c6f726967696e322e6578616d706c65","nonce":"f93a21c15f70a2e67f5efa2f9a5b504c8c8ef895c3f61f31f4abb3a1dc5e2b60","token_key_id":"ca572f8982a9ca248a3056186322d93ca147266121ddeb5632c07f1f71cd2708","token_authenticator_input":"0002f93a21c15f70a2e67f5efa2f9a5b504c8c8ef895c3f61f31f4abb3a1dc5e2b601f1d1cb68c578c37d20f2cf5cc192a5f58ee5621803d8d0b7fe97e770634673eca572f8982a9ca248a3056186322d93ca147266121ddeb5632c07f1f71cd2708"}]
