+++ tokens/type1/type1-issuance-test-vectors.json
@@ -1,0 +1,1 @@
+[{"skS":"39b0d04d3732459288fc5edb89bb02c2aa42e06709f201d6c518871d518114910bee3c919bed1bbffe3fc1b87d53240a","pkS":"02d45bf522425cdd2227d3f27d245d9d563008829252172d34e48469290c21da1a46d42ca38f7beabdf05c074aee1455bf","token_challenge":"0001000e6973737565722e6578616d706c65205de58a52fcdaef25ca3f65448d04e040fb1924e8264acfccfc6c5ad451d582b3000e6f726967696e2e6578616d706c65","nonce":"9f49c4ef01b37d0d518a3225a04e79c25fdfe2b1f7e7df7c75322c0e44674950","blind":"90356dc5029a1de88cad16f652302107492c25b2b627c0b8443c919effe9a5979834a08b8054397659c0be8c75432f5d","token_request":"0001f4033fbe5839b9f5747ce2f8074fd324fb15e501c8620168f0b2ad03d3685be2379c851d43fa9ab70cb8f31afa2678bc8c49","token_response":"0282c3f66149686bdb1e446994640e9ea6a051b34444e4e6843568a3cf51ce70cbb8ecd7e18905b1eba9d451ec7f156f95700d557ab8f5c902ad2c38d211ce1c712ded3e08313275b1d6ee70dcd9177d9425edaf79ef8bc163fda292e61aefeb27aa34e9b763c73b6e8677db0f225c8cf543c54ca3accef5af794f5e6dac72d89d5688cdf5372d7b33a7d193991a3e745a","token":"00019f49c4ef01b37d0d518a3225a04e79c25fdfe2b1f7e7df7c75322c0e44674950501370b494089dc462802af545e63809581ee6ef57890a12105c28368169514bf260d0792bf7f46c9866a6d37c3032d8714415f87f5f6903d7fb071e253be2f42d13f499899602c1b30be766dba722efd9d2120ce27fd5100570c8a60d55eebecf36beb4c102c0e091c1598ce538e381"},{"skS":"39efed331527cc4ddff9722ab5cd35aeafe7c27520b0cfa2eedbdc298dc3b12bc8298afcc46558af1e2eeacc5307d865","pkS":"038017e005904c6146b37109d6c2a72b95a183aaa9ed951b8d8fb1ed9033f68033284d175e7df89849475cd67a86bfbf4e","token_challenge":"0001000e6973737565722e6578616d706c6500000e6f726967696e2e6578616d706c65","nonce":"9a3fcef7cf25593e62772ca2cfde3314f763e6223599cf62819a1a3d2bdbd094","blind":"b748bf80e1d7f32b7f9b5a090624aa1defdcfc25efd9f048c4777171177402a5ecad001de20f27fb8cd7403b6f4d8e8a","token_request":"00013302671691d610d6ab13550b56695ef1e55152cccf71f032f6b7b056b7a23d8648403cb08c383dafd7de3c87a23c01735cf1","token_response":"0292c247815b8986f7617607d7a2ac19e36dbfc340e875e94d2b613584816c7268baf15c1a68a4b38f7394e9947cf0c74771e54bdadf244c37c23389653c79d0e463eea80f5e8ee9fa558c949df23d90d99c06d60a0550008c5e9f7628e751f5717ff320cd06b994c9b961675f8c950b55451ce225f02b5a631e80801a72ecd16f120957739ba026f23190720ad655fb47","token":"00019a3fcef7cf25593e62772ca2cfde3314f763e6223599cf62819a1a3d2bdbd094c994f7d5cdc2fb970b13d4e8eb6e6d8f9dcdaa65851fb091025dfe134bd5a62a116477bc9e1a205cca95d0c92335ca7a3e71063b2ac020bdd231c66097f12333142c52e960ba4e264ae79618bfe346334e7ab9065c6dcb2a7dfb38dfd0522a54f2ce32056ec6be5dd41c07d2f62f535d"},{"skS":"2b7709595b62b784f14946ae828f65e6caeba6eefe732c86e9ae50e818c055b3d7ca3a5f2beecaa859a62ff7199d35cc","pkS":"03a0de1bf3fd0a73384283b648884ba9fa5dee190f9d7ad4292c2fd49d8b4d64db674059df67f5bd7e626475c78934ae8d","token_challenge":"0001000e6973737565722e6578616d706c65000017666f6f2e6578616d706c652c6261722e6578616d706c65","nonce":"6cb636267732587fc924b9048ee24f7ba5df6487aceb12b707a496f478f8baa1","blind":"1fc10acd5107fbee03effdf4169a551847992b7b3fa210e7594afd68191edb2f3c5108ae10b190684e7532c116aefefc","token_request":"0001c802acc948b127f0e2c42463a93326ba1e313e80eb8704f5d0835959a7c41cace71c51485883705969e6c0c4da3c23f827fd","token_response":"032745b15c9393418c1330c2331b68dc8910912d0a634dd3e613646a3b6aaf9ff924e744c35295c50a9bddbc3d93b56b779a86b5bb5907a8a2a6dac6e643ea053c0b34abd475bee04eb42d67da80dc34af3f233880205be5da35b071bc43ad7e6246c59545cc9d2789d673840eeb94691e04ec2712f1d1339cc083183ca4a40c792e017cbc741bb0e97d885d61191cb42d","token":"00016cb636267732587fc924b9048ee24f7ba5df6487aceb12b707a496f478f8baa11949fd455872478ba87e2e6c513c3261cddbe57220581245e4c9c911dd1c0bb865785bff8f3cfe08cccbb3a7b8e41d23a172871be4828cc54582d87bc7cfc5c80b1d3921e801a39ba0d03ebc537db7991b54b0b851f6bdcc3f5045d7788196257847fa4241f6197af04794975b7a3482"},{"skS":"22e237b7b983d77474e4495aff2fc1e10422b1d955192e0fbf2b7b618fba625fcb94b599da9113da49c495a48fbf7f7f","pkS":"028cd68715caa20d19b2b20d017d6a0a42b9f2b0a47db65e5e763e23744fe14d74e374bbc93a2ec3970eb53c8aa765ee21","token_challenge":"0001000e6973737565722e6578616d706c65000000","nonce":"6871e48373742ba4e2b5544d5840c288b1961fdbee4e21d4837cf9d879309db0","blind":"5e5d161628107ce048938c6db38e20f15046247e1ad57ff355809772ef7ee9f4ae5244f342d2c27cd0bb3d1b7c7c7507","token_request":"0001a502702ca2eb1b58083bdb55c56a042e3791bb9194c7bfd5d5308cb989cd13599a104382640e9f197bfaf5e49ced519cf70b","token_response":"02bc134a08d7c6066e83ff13622e0684fdd4b4131d557d34b3fc13391ce728ddceac8c33cd3cf56cfe774a00f9e8ee73860c1246c8f51a4f83b40a7046675f2597ec8f9a5b8e0fc14226730fab35d022f469cbbdf306bdf0b90d2d4be288979bdede6a0df55b214776d324b26eb9bccb5ccef5232267550eb677fb16ecf9431c04d3110cb8e4cfea5eb0a512cbe3cb42a0","token":"00016871e48373742ba4e2b5544d5840c288b1961fdbee4e21d4837cf9d879309db0085cb06952044c7655b412ab7d484c97b97c48c79c568140b8d49a02ca47a9cfb0a5cfb861290c4dbd8fd9b60ee9b1a1a54cf47c98531fe253f1ed6d875de5a5590c9df9276b74cb4627f09122f55f4dc595c5f89420fcbe996bd20625e097aaae7d2aaa67588b5587d60f837c180e60"},{"skS":"46f3d4f562002b85ffcfdb4d06835fb9b2e24372861ecaa11357fd1f29f9ed26e44715549ccedeb39257f095110f0159","pkS":"02fbe9da0b7cabe3ec51c36c8487b10909142b59af030c728a5e87bb3b30f54c06415d22e03d9212bd3d9a17d5520d4d0f","token_challenge":"0001000e6973737565722e6578616d706c65205de58a52fcdaef25ca3f65448d04e040fb1924e8264acfccfc6c5ad451d582b30000","nonce":"2f12f86cd9391bd9e9b6caa7174e556081ffd2b32e9da2877303b54f1302ab2e","blind":"74bd5ee138202797653af239046580fc08205fe1aa3b00eb05514ad077f64bdfb18aa408635a8db294b91db1a07c78b8","token_request":"0001e102bc2fefbd549ea25a30ad4a1eebcdefee381c6a63ca15924bc8b8419760220780b580e0be920a31ff724d0fd092bf1414","token_response":"021494cb008719a61b9ad8baf68117784671353cd291fe4f8358ea228fdf16ca9b38fd16a509908f32aee3b60ba2e6b755bd552610d2b892acced62e7616e8c5a514a5fbe8aac1f7f71f5bedd0b8b8e63d039ccd47554a82a26860d0706f2811d6fa284f37a1685bc3842ebb978566016174974f8e76c087d705f55aeb83cfff1a135597b0fc819e8abe7d10fa48ce9aea","token":"00012f12f86cd9391bd9e9b6caa7174e556081ffd2b32e9da2877303b54f1302ab2ed4380df12a1727f4e2ca1ee0d7abea0d0fb1e9506507a4dd618f9b87e79f9f3521a7c9134d6722925bf622a994041cdb1b082cdf1309af32f0ce00ca1dab63e12e15d4d7426c594daf2bd90d1895387b9e6bf5670cab1503ed8d7349ed9fc788a12afb0f31d1b119f0c0604a5f946c95"}]
