+++ tokens/type3/client.go
@@ -1,0 +1,258 @@
+package type3
+
+import (
+	"crypto"
+	"crypto/elliptic"
+	"crypto/rand"
+	"crypto/rsa"
+	"crypto/sha256"
+	"crypto/sha512"
+
+	hpke "github.com/cisco/go-hpke"
+	"github.com/cloudflare/circl/blindsign"
+	"github.com/cloudflare/circl/blindsign/blindrsa"
+	"github.com/cloudflare/pat-go/ecdsa"
+	"github.com/cloudflare/pat-go/tokens"
+	"golang.org/x/crypto/cryptobyte"
+)
+
+type RateLimitedClient struct {
+	curve     elliptic.Curve
+	secretKey *ecdsa.PrivateKey
+}
+
+func NewRateLimitedClientFromSecret(secret []byte) RateLimitedClient {
+	curve := elliptic.P384()
+	secretKey, err := ecdsa.CreateKey(curve, secret)
+	if err != nil {
+		panic(err)
+	}
+
+	return RateLimitedClient{
+		curve:     elliptic.P384(),
+		secretKey: secretKey,
+	}
+}
+
+func padOriginName(originName string) []byte {
+	N := 31 - ((len(originName) - 1) % 32)
+	zeroes := make([]byte, N)
+	return append([]byte(originName), zeroes...)
+}
+
+func unpadOriginName(paddedOriginName []byte) string {
+	lastNonZero := len(paddedOriginName) - 1
+	for {
+		if lastNonZero < 0 {
+			// The plaintext was empty, so the input was the empty string
+			return ""
+		}
+		if paddedOriginName[lastNonZero] != 0x00 {
+			break
+		}
+		lastNonZero--
+	}
+	return string(paddedOriginName[0 : lastNonZero+1])
+}
+
+// https://ietf-wg-privacypass.github.io/draft-ietf-privacypass-rate-limit-tokens/draft-ietf-privacypass-rate-limit-tokens.html#name-encrypting-origin-token-req
+func encryptOriginTokenRequest(nameKey EncapKey, tokenKeyID uint8, blindedMessage []byte, requestKey []byte, originName string) ([]byte, []byte, []byte, error) {
+	issuerKeyEnc := nameKey.Marshal()
+	issuerKeyID := sha256.Sum256(issuerKeyEnc)
+
+	enc, context, err := hpke.SetupBaseS(nameKey.suite, rand.Reader, nameKey.publicKey, []byte("TokenRequest"))
+	if err != nil {
+		return nil, nil, nil, err
+	}
+
+	b := cryptobyte.NewBuilder(nil)
+	b.AddUint8(nameKey.id)
+	b.AddUint16(uint16(nameKey.suite.KEM.ID()))
+	b.AddUint16(uint16(nameKey.suite.KDF.ID()))
+	b.AddUint16(uint16(nameKey.suite.AEAD.ID()))
+	b.AddUint16(RateLimitedTokenType)
+	b.AddBytes(requestKey)
+	b.AddBytes(issuerKeyID[:])
+
+	tokenRequest := InnerTokenRequest{
+		blindedMsg:   blindedMessage,
+		tokenKeyId:   tokenKeyID,
+		paddedOrigin: padOriginName(originName),
+	}
+	input := tokenRequest.Marshal()
+
+	aad := b.BytesOrPanic()
+
+	ct := context.Seal(aad, input)
+	encryptedTokenRequest := append(enc, ct...)
+	secret := context.Export([]byte("TokenResponse"), nameKey.suite.AEAD.KeySize())
+
+	return issuerKeyID[:], encryptedTokenRequest, secret, nil
+}
+
+type RateLimitedTokenRequestState struct {
+	tokenInput        []byte
+	clientKey         []byte
+	blindedRequestKey []byte
+	request           *RateLimitedTokenRequest
+	encapSecret       []byte
+	encapEnc          []byte
+	nameKey           EncapKey
+	verificationKey   *rsa.PublicKey
+	verifier          blindsign.VerifierState
+}
+
+func (s RateLimitedTokenRequestState) Request() *RateLimitedTokenRequest {
+	return s.request
+}
+
+func (s RateLimitedTokenRequestState) RequestKey() []byte {
+	return s.blindedRequestKey
+}
+
+func (s RateLimitedTokenRequestState) ClientKey() []byte {
+	return s.clientKey
+}
+
+// https://ietf-wg-privacypass.github.io/draft-ietf-privacypass-rate-limit-tokens/draft-ietf-privacypass-rate-limit-tokens.html#name-attester-to-client-response
+func (s RateLimitedTokenRequestState) FinalizeToken(encryptedtokenResponse []byte) (tokens.Token, error) {
+	// response_nonce = random(max(Nn, Nk)), taken from the encapsualted response
+	responseNonceLen := max(s.nameKey.suite.AEAD.KeySize(), s.nameKey.suite.AEAD.NonceSize())
+
+	// salt = concat(enc, response_nonce)
+	salt := append(s.encapEnc, encryptedtokenResponse[:responseNonceLen]...)
+
+	// prk = Extract(salt, secret)
+	prk := s.nameKey.suite.KDF.Extract(salt, s.encapSecret)
+
+	// aead_key = Expand(prk, "key", Nk)
+	key := s.nameKey.suite.KDF.Expand(prk, []byte(labelResponseKey), s.nameKey.suite.AEAD.KeySize())
+
+	// aead_nonce = Expand(prk, "nonce", Nn)
+	nonce := s.nameKey.suite.KDF.Expand(prk, []byte(labelResponseNonce), s.nameKey.suite.AEAD.NonceSize())
+
+	cipher, err := s.nameKey.suite.AEAD.New(key)
+	if err != nil {
+		return tokens.Token{}, err
+	}
+
+	// reponse, error = Open(aead_key, aead_nonce, "", ct)
+	blindSignature, err := cipher.Open(nil, nonce, encryptedtokenResponse[responseNonceLen:], nil)
+	if err != nil {
+		return tokens.Token{}, err
+	}
+
+	signature, err := s.verifier.Finalize(blindSignature)
+	if err != nil {
+		return tokens.Token{}, err
+	}
+
+	tokenData := append(s.tokenInput, signature...)
+	token, err := UnmarshalToken(tokenData)
+	if err != nil {
+		return tokens.Token{}, err
+	}
+
+	// Sanity check: verify the token signature
+	hash := sha512.New384()
+	_, err = hash.Write(token.AuthenticatorInput())
+	if err != nil {
+		return tokens.Token{}, err
+	}
+	digest := hash.Sum(nil)
+
+	err = rsa.VerifyPSS(s.verificationKey, crypto.SHA384, digest, token.Authenticator, &rsa.PSSOptions{
+		Hash:       crypto.SHA384,
+		SaltLength: crypto.SHA384.Size(),
+	})
+	if err != nil {
+		return tokens.Token{}, err
+	}
+
+	return token, nil
+}
+
+// https://ietf-wg-privacypass.github.io/draft-ietf-privacypass-rate-limit-tokens/draft-ietf-privacypass-rate-limit-tokens.html#name-client-to-attester-request
+func (c RateLimitedClient) CreateTokenRequest(challenge, nonce, blindKeyEnc []byte, tokenKeyID []byte, tokenKey *rsa.PublicKey, originName string, nameKey EncapKey) (RateLimitedTokenRequestState, error) {
+	blindKey, err := ecdsa.CreateKey(c.curve, blindKeyEnc)
+	if err != nil {
+		return RateLimitedTokenRequestState{}, err
+	}
+
+	clientKeyEnc := elliptic.MarshalCompressed(c.curve, c.secretKey.PublicKey.X, c.secretKey.PublicKey.Y)
+
+	b := cryptobyte.NewBuilder(nil)
+	b.AddUint16(RateLimitedTokenType)
+	b.AddBytes([]byte("ClientBlind"))
+	ctx := b.BytesOrPanic()
+	blindedPublicKey, err := ecdsa.BlindPublicKeyWithContext(c.curve, &c.secretKey.PublicKey, blindKey, ctx)
+	if err != nil {
+		return RateLimitedTokenRequestState{}, err
+	}
+	blindedPublicKeyEnc := elliptic.MarshalCompressed(c.curve, blindedPublicKey.X, blindedPublicKey.Y)
+
+	verifier := blindrsa.NewRSAVerifier(tokenKey, crypto.SHA384)
+
+	context := sha256.Sum256(challenge)
+	token := tokens.Token{
+		TokenType:     RateLimitedTokenType,
+		Nonce:         nonce,
+		Context:       context[:],
+		KeyID:         tokenKeyID,
+		Authenticator: nil, // No signature computed yet
+	}
+	tokenInput := token.AuthenticatorInput()
+	blindedMessage, verifierState, err := verifier.Blind(rand.Reader, tokenInput)
+	if err != nil {
+		return RateLimitedTokenRequestState{}, err
+	}
+
+	nameKeyID, encryptedTokenRequest, secret, err := encryptOriginTokenRequest(nameKey, tokenKeyID[0], blindedMessage, blindedPublicKeyEnc, originName)
+	if err != nil {
+		return RateLimitedTokenRequestState{}, err
+	}
+
+	b = cryptobyte.NewBuilder(nil)
+	b.AddUint16(RateLimitedTokenType)
+	b.AddBytes(blindedPublicKeyEnc)
+	b.AddBytes(nameKeyID)
+	b.AddUint16LengthPrefixed(func(b *cryptobyte.Builder) {
+		b.AddBytes(encryptedTokenRequest)
+	})
+	message := b.BytesOrPanic()
+
+	hash := sha512.New384()
+	hash.Write(message)
+	digest := hash.Sum(nil)
+
+	r, s, err := ecdsa.BlindKeySignWithContext(rand.Reader, c.secretKey, blindKey, digest, ctx)
+	if err != nil {
+		return RateLimitedTokenRequestState{}, err
+	}
+	scalarLen := (c.curve.Params().Params().BitSize + 7) / 8
+	rEnc := make([]byte, scalarLen)
+	sEnc := make([]byte, scalarLen)
+	r.FillBytes(rEnc)
+	s.FillBytes(sEnc)
+	signature := append(rEnc, sEnc...)
+
+	request := &RateLimitedTokenRequest{
+		RequestKey:            blindedPublicKeyEnc,
+		NameKeyID:             nameKeyID,
+		EncryptedTokenRequest: encryptedTokenRequest,
+		Signature:             signature,
+	}
+
+	requestState := RateLimitedTokenRequestState{
+		tokenInput:      tokenInput,
+		clientKey:       clientKeyEnc,
+		request:         request,
+		encapSecret:     secret,
+		encapEnc:        encryptedTokenRequest[0:nameKey.suite.KEM.PublicKeySize()],
+		nameKey:         nameKey,
+		verifier:        verifierState,
+		verificationKey: tokenKey,
+	}
+
+	return requestState, nil
+}
