+++ tokens/type3/type3-anon-origin-id-test-vectors.json
@@ -1,0 +1,1 @@
+[{"sk_sign":"0952a4aa40c3df8e9903d8b0cd7a527201db8611dd060b3b534f7ba4042573f3860dd0a4781cff2ae815f077c4cd746f","pk_sign":"02caf8455d3f97eae5040cb02c0598776e7ebc0f7c7271b4b17af5c26dd80a102e90548ef7cbfd7d53195a8a9c61ccbcc2","sk_origin":"cb9effb92af3ac8098896d4a5b319bc447a012dfb64a90a3e7ae42d34b701ee897710726eb9326065eea61139309478a","request_blind":"b47a7d23a6dab24a924d5ae37621547c3de28a2e9df971e1eee3027ef82385fd0e3f5045ac7b61735a22cb1bd57ffb1a","request_key":"027acce859efe602f8d22fb7c4fef2af15d4e123bf7496286efa569163be06c482db1b22c7ffcfbe8ec88e8877b2dc8ec9","index_key":"0388d960e0455fea77c9453689cedbde815b80ab21269f38a216399ffb57bcb5b000ee89a933e6fd15e3a7e5f7d797932f","anon_issuer_origin_id":"fce2702d46ea48bdd3dc8d9c4e783a0610d153c64b324f631a49b49e79f11f686701d8817c19fb97c199c1c3b780c0de"}]
