+++ tokens/type3/type3-ecdsa-blinding-test-vectors.json
@@ -1,0 +1,1 @@
+[{"Curve":"P-384","Hash":"SHA-384","skS":"a817e2f524707db3badf31467b9da847fcd8a7eee70c1f6905dcd494c72cbe03a4442a9e30ae043a1623d23a3383b61d","pkS":"03b5580e94a5d97b699975b13f20fc5a83485f8afa19e267fbbd564049bb46d25830e107e9b98077b1b3a0aaf23d279423","bk":"03f40dada41ce442501b6bab944c7bcdacb133a7f25e7ab570f720f6497a85eb01ef09b68773d83f2cea2e780355b097","pkR":"029073f469a936dc20d7dba22a58b2f2e6a201eaf098612923367d86798fc07b6dc25682b8a4cefa1e8aff3f90688f9cf0","message":"68656c6c6f20776f726c64","context":"","signature":"72ddb8cd4aeccaa08fb8f500c939c9698c5e092c8481bcf0e4f8791bba3f76fd61bccc58c719c66f320560391df990175dce37c8580771b50572801f2ef8753b2706d0b8c021f7b10c6a794dc1ebf2717fe054ade947e6cd6900483b0212d7ca"},{"Curve":"P-384","Hash":"SHA-384","skS":"d745530c3f6ad80e27934983eb403fecb5498a0b95cd3f320bcc6440fbf2737e752423299c71746e41c9ef53926481c9","pkS":"02f9f13e3c7b15164334741fe123e7c58944959e61e3504eb8dcb37a78a4cef2a9d2c1ad0bd39f6de6098d3191e0cf3404","bk":"5ca030a7ea370a832a721bea1466eceff5a17ccb557a94ee65d994a10125eeef236a61eb3ad27a6632dd40664b0cdf70","pkR":"025661c063757acbf99d09a695c74cc6195e1a775c4c3a0896c2799d7580b2737589d49c81e208531efd1c1de2f293835f","message":"68656c6c6f20776f726c64","context":"65c6014d3f36eed6f525c80f2bdf856a2776a23ba26b900bfdbd7e7d6a3b32fe","signature":"c913493c423df166fe5697be7aafd3c9c5b8a699b9e42b2662eefe2a48945ea4df6b921da8dae2ba47a98acf54625946f207fc64b19c4051ba147185dc312911db300544105ba8d3e8205c5276b62a7f9acae14539c5bbd4faa9bac7d96c0bdc"}]
