+++ tokens/type3/type3-ed25519-blinding-test-vectors.json
@@ -1,0 +1,1 @@
+[{"skS":"d9e696a25f28b52d283ce8f4ad22f306c151137a9a4f734dfaf748312f9421b5","pkS":"1c9122175691cf4dd70f65e7f0035b5aaf6017cbe4a15d01281d878e4c152fd5","bk":"a3a646b88958c3c62e9143ffbc8a0169e67935e65139117512958bf57e7e3dc4","pkB":"","pkR":"8563b459df57a4ef6eabb675171096d9eb758e7fc950779174fe6c9c6ab56d1b","message":"68656c6c6f20776f726c64","context":"","signature":"54673c2b9d34a47dc2c3e7980ed9400c104b71eba24331650c3331c9ef1eed53b8940cfb8bfe3a2a360d3bf9dd3625c333514dfa86054bdbf88152f2d79fed0c"},{"skS":"899a49673f17baa67e5ec07942da38d03265d355ac6ccb24edb8e13c41d09d85","pkS":"d61962c6ceee27046aa2c4300168dcd4dcb4b1136c141c67186b8dca991f986c","bk":"0000000000000000000000000000000000000000000000000000000000000000","pkB":"","pkR":"c0e39e3251c3b2fe62e8c142175c325ddd5d72312d489b431eae5ec1c6d2d199","message":"68656c6c6f20776f726c64","context":"","signature":"0622f0be9f53259700bde5c46aeea09e3b4a31c5c96831a65c28c7330658fbc2104679ae41170ec2e695c63c5234f37b3be34431606a6c6b74a2a00335c5d00d"},{"skS":"89e7ffccd89d752fba82c894b5698d94e01c243b3c17130d8ad4eee32e79f84c","pkS":"b20736c6b81bf80b3eb572aa0e94d286e83529b7f87aa4a46c0994aaabb990f7","bk":"fcc514425b2ef67527126f8f825dfe566af0fc810a0781fd1d97da240b4378ac","pkB":"","pkR":"c2819d1ec69959860195504ee6c571709fbabb0d4acb7154b51a21deecea6eb0","message":"68656c6c6f20776f726c64","context":"22004056b5156da84ee77cdf1294fecd3d5660f859d35bad2de4a4632fad84c9","signature":"397acc3f89d3ff24f7b9a9045395647eae98fde7b90fe5969547bc695a4de66a04dc28cea14cbe310baf3abd613edc50c46f64cc8a2e48c0b25662e1ae570f04"},{"skS":"3764ed6209bce77a67c0c1adac8cf35a992f6c601bb1709b89b620bccbde4cf2","pkS":"54b6eed6784bd5898cf0823a53a8fc2aceb484755bc0df95dc904d62d3818aa0","bk":"0000000000000000000000000000000000000000000000000000000000000000","pkB":"","pkR":"529966d5df826df27d65bf42a315e436d0dfd0e907cbbf7bd4da3b4e94d16517","message":"68656c6c6f20776f726c64","context":"a9afa52c87a24558c88425eae85afb9e3a4e2d26f6b2f128bffd195639868293","signature":"a6dcca1196138167bbd02da51b7a70eecb61301f4f7903cc32d69fa68ab2fd04aa0d2c922ed7796a9f8f7dd304d5534354a26c6ddc0da56db4b9be9822ff1502"}]
