+++ tokens/type3/type3-origin-encryption-test-vectors.json
@@ -1,0 +1,1 @@
+[{"kem_id":32,"kdf_id":1,"aead_id":1,"issuer_encap_key_seed":"9476057d4168bea2d68438550455b6b4d2ac6bb99f6e2d1ef2c4d3b9e2fa73b5","issuer_encap_key":"0100202407491be7660db25ea0d8c305431ab85622e2d3f49492a3c9fe7ec91d1ebf6b00010001","token_type":3,"issuer_encap_key_id":"6a8a37ff3723048c2f831ab0892725dcab26e306aeee0f86d7e8441aabe2aa42","request_key":"bfd07a4a4f54003a2092e7c957508e37a2655e91d2249a9fd4d4c4ea511dbf29271581e41c0eb99918a225ddd839d90193","token_key_id":35,"blinded_msg":"cdc3891d2cb53df7dacd18ddcf4cd2060e1afe2b4e496d7f3ddb5a4fa7e490360877096334fd7e2262f3697c7f40db758f13cf4ac4caf7d452f7de947c214c9684c436c226f24581d30a5115885ae888068d8a67b66911ef888cfba856962b6f6a14a1a23ca08301c02e5e4b9a46963bf7f822e0385f7ddbfc4d40ce26d3c707d726e59858b812a46493f46dcdb1c73e667c9ad746da40bd725937116fdca6cb5e03c5a2586b1745b72316ddafa3583a55f9442eb949277ce4183c15549b9c35ca3c4554792e0d195cec5f391fbf0cec1bb4a5f93c5e0205110abe64bd6c9dc4425a04a9b516d770cc35f0b957a55231e08f0b5cb0dcc86bf2863282f625d844","origin_name":"746573742e6578616d706c65","encap_secret":"2c4605e0c6700b101e11de630e4fc088","encrypted_token_request":"c6642c1aea2f047e626d46603706242849f896c50c67a17d1a46a9af6823697ee3d323542aaa6e5e8e8663f4d2352e0a7e3a04328ff080040b10ccc50641b4ccc1d45404ff3ab7bf107fd6fcaa15255c97039d3161f0017fb5384c33f06a354f8776836091b5c51477e994930694e21a07231d7fd892ff8c4f6b6c2ed7a75b5b4e90191e61111ce27c6df3c6a68b985175745ec17f93102010444009e2905c327a104d5716c8b7641226047ea337b11c06ecabbfc1a57b24b5a0684d9734e5fb8329f2e3543245571c3ddf8c94d7cf6c58593b1350c804424c814fd281078b537bdb6d17567e455ab41163694e8f4a3bb0ad6946743d3e3c7aeb99fd6920302ad296ce2c7dc50e2e15bfdb195c838332cdf7fa638e8e716ed50b0d4c4686b16348413f4e3795899c697fdf56267e2029bf77879510ef2a1c983ae9f4036406c5ed54f33d334c21e93041e4bf271122f90563bf"}]
